src/load-generator/requirements.txt
backoff==2.2.1
brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
configargparse==1.7
deprecated==1.2.14
flask==3.0.0
flask-basicauth==0.2.0
flask-cors==4.0.0
gevent==23.9.1
geventhttpclient==2.0.11
googleapis-common-protos==1.61.0
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
jinja2==3.1.2
locust==2.18.2
locust_plugins==3.4.0
markupsafe==2.1.3
msgpack==1.0.7
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-jinja2==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-system-metrics==0.46b0
opentelemetry-instrumentation-urllib3==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
openfeature-sdk==0.7.0
openfeature-provider-flagd==0.1.5
openfeature-hooks-opentelemetry==0.1.3
protobuf==4.25.2
psutil==5.9.6
pyzmq==25.1.1
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
setuptools==75.1.0
typing-extensions==4.8.0
urllib3==2.0.7
werkzeug==3.0.1
wrapt==1.16.0
zope-event==5.0
zope-interface==6.1
python-json-logger==2.0.7
